# perf: Implement compiler sharding