--- a/.gitignore
+++ b/.gitignore
@@ -12,6 +12,7 @@ bin
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+*.png
 
 # Kubernetes Generated files - skip generated files, except for vendored files