@willbeason 2022-03-03
I'm taking benchmarks now. See data here: https://docs.google.com/spreadsheets/d/1WYLMDdr9w9QwF9NGYvF1bzrYrKwF52Efj4bl7jaGMqA/edit?usp=sharing

I'll post a summary once I've taken all of the benchmarks for AddTemplate/AddConstraint/Review