@willbeason 2022-03-03
Benchmarks are in the above spreadsheet now, but I need to write about how to interpret them. Those will be coming tomorrow.